Use Nectar gerrit (start new release branch)
loci does not have release branches, but we do not want to keep
committing to master as it diverges from upstream and is confusing.

We have decided to cut nectar/* branches for now from upstream at
release time. This will ensure that master always tracks upstream, and
we can cherry-pick changes to each release branch

Change-Id: I770630747a69b12329b27e997fdd1f2614ff3a29